WIP: split place reassignment e2e coverage (checkpoint)
